+++ MustafaevApplication2/Station.h
@@ -4,16 +4,16 @@ private:
 	std::string name;
 	int countWorkshop;
 	int countActiveWorkshop;
-	int efficiency;
+	double efficiency;
 
 public:
 	Station();
 	void setName(std::string name);
 	void setCountWorkshop(int countWorkshop);
 	void setCountActiveWorkshop(int countActiveWorkshop);
-	void setEfficiency(int efficiency);
+	void setEfficiency(double efficiency);
 	std::string getName() const;
 	int getCountWorkshop() const;
 	int getCountActiveWorkshop() const;
-	int getEfficiency() const;
+	double getEfficiency() const;
 };
